Update README.md
Neuromorphic neural network with 3D and dual-hemisphere processing Patent PCT/IR2025/050026 Protected codes and initial network setup are not yet complete. This network will be programmed in 4 models. The first model is the first preprocessing layer, dual-hemisphere hidden layers, and a corpus callosum simulation layer for separate and coordinated processing between the two hemispheres, the second model has 4 hemispheres, exactly the brain model with two inputs, and eagle bird vision with a processing method without using convolutional methods. All of these are under the ownership rights of this invention. As well as piezoelectric memory sensors and smart circuits and physical intelligence that include nanoscale transistors from carbon nanotubes, complex pattern processing, energy efficiency, high speed, interpretability, real-time decision-making in robotics, minimal reduction of illusions and delusions in artificial intelligence, lower cost of calculations, rapid model training, self-organization, and soon under the license of this network to companies in Europe and the United States, the ability will be transferred with permission to manufacture new generation chips. According to the provisions of the invention, this invention does not and does not have any military, defense, or drone applications, and will be subject to international prosecution if used. 70% of the ownership rights of this invention and the rights arising from it have been donated to associations supporting orphaned children and women heads of households around the world by the inventor Nader Maleki.